backend/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
Markdown==3.7
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.0
sqlparse==0.5.1
typing_extensions==4.12.2
